requirements.txt
certifi==2025.1.31
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
dnspython==2.7.0
motor==3.7.0
pymongo==4.11.1
requests==2.32.3
opencv-python==4.11.0.86
face-recognition==1.3.0
face_recognition_models==0.3.0
sentence-transformers==3.4.1
transformers==4.48.2
spacy
open_clip_torch==2.31.0 
torch==2.6.0
open_clip_torch==2.31.0 
anyio==4.8.0 
distro==1.9.0 
groq==0.18.0 
h11==0.14.0 
httpcore==1.0.7 
httpx==0.28.1 
sniffio==1.3.1
loguru==0.7.3
boto3==1.37.2
botocore==1.37.2
jmespath==1.0.1 
s3transfer==0.11.3
dotenv==0.9.9 
python-dotenv==1.0.1